# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0",
    "title": "PlooralApiRestTest",
    "description": "API REST Test Plooral"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "LocalTest"
    },
    {
      "url": "http://ec2-54-94-2-165.sa-east-1.compute.amazonaws.com",
      "description": "LocalProd"
    }
  ],
...
  "paths": {
    "/v1/jobs/": {
      "post": {
        "tags": [
          "jobs"
        ],
        "description": "Endpoint to create a job posting draft.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "177e8dcf-02d9-47c5-9671-8d3369c0ec56"
                    },
                    "status": {
                      "type": "string",
                      "example": "draft"
                    },
                    "created_at": {
                      "type": "string",
                      "example": "2024-02-16T02:18:41.517Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "example": "2024-02-16T02:18:41.517Z"
                    },
                    "company_id": {
                      "type": "string",
                      "example": "38028bfc-a6fe-40cb-b8d3-9529541fd376"
                    },
                    "title": {
                      "type": "string",
                      "example": "Dev C++"
                    },
                    "description": {
                      "type": "string",
                      "example": "c plus plus"
                    },
                    "location": {
                      "type": "string",
                      "example": "Remote"
                    },
                    "notes": {
                      "type": "string",
                      "example": "A"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2024-02-16T02:18:41.517Z"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2024-02-16T02:18:41.517Z"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "177e8dcf-02d9-47c5-9671-8d3369c0ec56"
                    },
                    "status": {
                      "type": "string",
                      "example": "draft"
                    },
                    "created_at": {
                      "type": "string",
                      "example": "2024-02-16T02:18:41.517Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "example": "2024-02-16T02:18:41.517Z"
                    },
                    "company_id": {
                      "type": "string",
                      "example": "38028bfc-a6fe-40cb-b8d3-9529541fd376"
                    },
                    "title": {
                      "type": "string",
                      "example": "Dev C++"
                    },
                    "description": {
                      "type": "string",
                      "example": "c plus plus"
                    },
                    "location": {
                      "type": "string",
                      "example": "Remote"
                    },
                    "notes": {
                      "type": "string",
                      "example": "A"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2024-02-16T02:18:41.517Z"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2024-02-16T02:18:41.517Z"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "BAD_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "'company_id is not found"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "'company_id is not found"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERROR"
                    },
                    "status": {
                      "type": "number",
                      "example": 500
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERROR"
                    },
                    "status": {
                      "type": "number",
                      "example": 500
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "company_id": {
                    "example": "any"
                  },
                  "title": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  },
                  "notes": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/publish": {
      "put": {
        "tags": [
          "jobs"
        ],
        "description": "Endpoint to create a publish a job posting draft.",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Job is published"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Job is published"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "BAD_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Job Status draft!"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Job Status draft!"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERROR"
                    },
                    "status": {
                      "type": "number",
                      "example": 500
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERROR"
                    },
                    "status": {
                      "type": "number",
                      "example": 500
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      }
    },
...
    "/v1/jobs/{job_id}/archive": {
      "put": {
        "tags": [
          "jobs"
        ],
        "description": "Endpoint to archive an active job posting.",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "08c04a8d-94b4-4136-ad4c-db9c73eb7197"
                    },
                    "company_id": {
                      "type": "string",
                      "example": "2ee5582c-b34c-4a48-af9f-3f2c4c5d4df4"
                    },
                    "title": {
                      "type": "string",
                      "example": "Dev C++"
                    },
                    "description": {
                      "type": "string",
                      "example": "C++"
                    },
                    "location": {
                      "type": "string",
                      "example": "Remote"
                    },
                    "notes": {
                      "type": "string",
                      "example": "A"
                    },
                    "status": {
                      "type": "string",
                      "example": "archived"
                    },
                    "created_at": {
                      "type": "string",
                      "example": "2024-02-16T14:30:18.585Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "example": "2024-02-16T14:30:18.585Z"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2024-02-16T14:30:18.585Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2024-02-16T14:37:50.746Z"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "08c04a8d-94b4-4136-ad4c-db9c73eb7197"
                    },
                    "company_id": {
                      "type": "string",
                      "example": "2ee5582c-b34c-4a48-af9f-3f2c4c5d4df4"
                    },
                    "title": {
                      "type": "string",
                      "example": "Dev C++"
                    },
                    "description": {
                      "type": "string",
                      "example": "C++"
                    },
                    "location": {
                      "type": "string",
                      "example": "Remote"
                    },
                    "notes": {
                      "type": "string",
                      "example": "A"
                    },
                    "status": {
                      "type": "string",
                      "example": "archived"
                    },
                    "created_at": {
                      "type": "string",
                      "example": "2024-02-16T14:30:18.585Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "example": "2024-02-16T14:30:18.585Z"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2024-02-16T14:30:18.585Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2024-02-16T14:37:50.746Z"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "BAD_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Job Status is archived!"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Job Status is archived!"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERROR"
                    },
                    "status": {
                      "type": "number",
                      "example": 500
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERROR"
                    },
                    "status": {
                      "type": "number",
                      "example": 500
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Token access"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Token access"
      }
    }
  }
}